move all rf_diffusion specific code to rf_diffusion
