+++ .pyright_exclude
@@ -1,0 +1,136 @@
+ipd/crud/__init__.py
+ipd/crud/backend.py
+ipd/crud/crudcli.py
+ipd/crud/frontend.py
+ipd/crud/server.py
+ipd/fit/__init__.py
+ipd/fit/qcp_rms.py
+ipd/fit/qcp_rms_cuda.py
+ipd/fit/qcp_scan.py
+ipd/homog/__init__.py
+ipd/homog/hgeom.py
+ipd/homog/pointgen.py
+ipd/homog/quat.py
+ipd/homog/thgeom.py
+ipd/homog/util.py
+ipd/pdb/__init__.py
+ipd/pdb/bbutil.py
+ipd/pdb/chemical.py
+ipd/pdb/pdbdump.py
+ipd/pdb/pdbfile.py
+ipd/pdb/pdbmeta.py
+ipd/pdb/pdbread.py
+ipd/pdb/pisces.py
+ipd/pdb/viz_pdb.py
+ipd/samp/__init__.py
+ipd/samp/sampling.py
+ipd/samp/sampling_cuda.py
+ipd/samp/tip_atom_placer.py
+ipd/sym/old/__init__.py
+ipd/sym/old/asufit.py
+ipd/sym/old/asugrid.py
+ipd/sym/old/asuslide.py
+ipd/sym/old/diffusion_sym.py
+ipd/sym/old/oldtests/__init__.py
+ipd/sym/old/oldtests/oldtest_asufit.py
+ipd/sym/old/oldtests/oldtest_asugrid.py
+ipd/sym/old/oldtests/oldtest_asuslide.py
+ipd/sym/old/oldtests/oldtest_diffusion_sym.py
+ipd/sym/old/oldtests/oldtest_pymol_symgen.py
+ipd/sym/old/oldtests/oldtest_rosetta_symdef.py
+ipd/sym/old/oldtests/oldtest_symfit.py
+ipd/sym/old/pymol_symgen.py
+ipd/sym/old/pymol_xyzmath.py
+ipd/sym/old/rosetta_symdef.py
+ipd/sym/old/symfit.py
+ipd/sym/xtal/SymElem.py
+ipd/sym/xtal/__init__.py
+ipd/sym/xtal/spacegroup.py
+ipd/sym/xtal/spacegroup_contacts.py
+ipd/sym/xtal/spacegroup_data.py
+ipd/sym/xtal/spacegroup_deriveddata.py
+ipd/sym/xtal/spacegroup_frames.py
+ipd/sym/xtal/spacegroup_symelems.py
+ipd/sym/xtal/spacegroup_util.py
+ipd/sym/xtal/xtalcls.py
+ipd/sym/xtal/xtalfit.py
+ipd/sym/xtal/xtalinfo.py
+ipd/tests/__init__.py
+ipd/tests/bvh/test_bvh_numba.py
+ipd/tests/check_test_results.py
+ipd/tests/conftest.py
+ipd/tests/crud/test_crud.py
+ipd/tests/cuda/test_cudafunc.py
+ipd/tests/dev/code/test_gentest.py
+ipd/tests/dev/observer/test_dynamic_parameters.py
+ipd/tests/dev/observer/test_observer.py
+ipd/tests/dev/test_bunch.py
+ipd/tests/dev/test_misc.py
+ipd/tests/dev/test_project_config.py
+ipd/tests/dev/test_state.py
+ipd/tests/dev/test_timer.py
+ipd/tests/fit/test_qcp_rms.py
+ipd/tests/fit/test_qcp_scan.py
+ipd/tests/fixtures.py
+ipd/tests/homog/__init__.py
+ipd/tests/homog/test_homog.py
+ipd/tests/homog/test_homog_misc.py
+ipd/tests/homog/test_quat.py
+ipd/tests/homog/test_thgeom.py
+ipd/tests/maintest.py
+ipd/tests/motif/test_motif_manager.py
+ipd/tests/pdb/test_pdbdat_build.py
+ipd/tests/pdb/test_pdbdump.py
+ipd/tests/pdb/test_pdbfile.py
+ipd/tests/pdb/test_pdbread.py
+ipd/tests/samp/test_sampling.py
+ipd/tests/samp/test_tip_atom_placer.py
+ipd/tests/sym/__init__.py
+ipd/tests/sym/helix/test_helix.py
+ipd/tests/sym/high_t/test_high_t_from_file.py
+ipd/tests/sym/high_t/test_pseudo_t.py
+ipd/tests/sym/sym_test_util.py
+ipd/tests/sym/test_guess_symmetry.py
+ipd/tests/sym/test_ipd_sym_manager.py
+ipd/tests/sym/test_permutations.py
+ipd/tests/sym/test_sym.py
+ipd/tests/sym/test_sym_adapt.py
+ipd/tests/sym/test_sym_check.py
+ipd/tests/sym/test_sym_color.py
+ipd/tests/sym/test_sym_fitting.py
+ipd/tests/sym/test_sym_manager.py
+ipd/tests/sym/test_sym_options.py
+ipd/tests/sym/test_sym_slice.py
+ipd/tests/sym/xtal/__init__.py
+ipd/tests/sym/xtal/test_SymElem.py
+ipd/tests/sym/xtal/test_spacegroup.py
+ipd/tests/sym/xtal/test_spacegroup_contacts.py
+ipd/tests/sym/xtal/test_spacegroup_data.py
+ipd/tests/sym/xtal/test_spacegroup_pymol.py
+ipd/tests/sym/xtal/test_spacegroup_symelems.py
+ipd/tests/sym/xtal/test_xtalcls.py
+ipd/tests/sym/xtal/test_xtalfit.py
+ipd/tests/test_pip_install.py
+ipd/tests/tools/ci/test_citool.py
+ipd/tests/tools/test_ipdtool.py
+ipd/tests/tools/test_run_tests_on_file.py
+ipd/tests/viz/test_plot.py
+ipd/tests/viz/test_pymol_viz.py
+ipd/tests/voxel/test_voxdock.py
+ipd/tests/voxel/test_voxel.py
+ipd/viz/__init__.py
+ipd/viz/plot.py
+ipd/viz/primitives.py
+ipd/viz/pymol_cgo.py
+ipd/viz/pymol_pose.py
+ipd/viz/pymol_viz.py
+ipd/viz/viz_bond.py
+ipd/viz/viz_deco.py
+ipd/viz/viz_helix.py
+ipd/viz/viz_rigidbody.py
+ipd/viz/viz_xtal.py
+ipd/voxel/__init__.py
+ipd/voxel/voxdock.py
+ipd/voxel/voxel.py
+ipd/voxel/voxel_cuda.py
+ipd/voxel/voxel_viz.py
