+++ git_pre_commit.sh
@@ -28,13 +28,13 @@ if [ -f .pyright_hash_last_commit ]; then
     cmd="PYTHONPATH=$ipd python -m ipd code pyright $src --hashfile '.pyright_hash_last_commit'"
     echo $cmd
     eval $cmd
-    git add .pyright_hash_last_commit
     if [ $? == 0 ]; then
         echo pyright pass
     else
         echo pyright fail
         exit 1
     fi
+    git add .pyright_hash_last_commit
 fi
 
 cmd="PYTHONPATH=$ipd python -m ipd code yapf $src"
