+++ ipd/__init__.py
@@ -15,7 +15,7 @@ motif = lazyimport('ipd.motif')
 pdb = lazyimport('ipd.pdb')
 qt = lazyimport('ipd.dev.qt')
 samp = lazyimport('ipd.samp')
-sieve = lazyimport('ipd.sieve')
+protocol = lazyimport('ipd.protocol')
 sym = lazyimport('ipd.sym')
 tests = lazyimport('ipd.tests')
 tools = lazyimport('ipd.tools')
