--- ipd/sieve/__init__.py
@@ -1,2 +1,0 @@
-from ipd.sieve.sieve import *
-from ipd.sieve import basic_sieves as basic_sieves
