+++ ipd/tests/sym/test_sym_color.py
